+++ CHANGELOG.md
@@ -39,6 +39,10 @@ request adding CHANGELOG notes for breaking (!) changes and possibly other secti
 
 ### Deprecations
 
+* The property `polaris.active-roles-provider.type` is deprecated in favor of
+  `polaris.authentication.active-roles-provider.type`. The old property is still supported, but will be removed in a
+  future release.
+
 ### Fixes
 
 ### Commits